[#28] Feat: 로그아웃/유저 검증 훅 생성 및  로그인 방식 파이어베이스 auth로 변경
