+++ src/components/Login/useAuthState.jsx
@@ -1,0 +1,19 @@
+import { useEffect, useState } from 'react'
+import { onAuthStateChanged } from 'firebase/auth'
+import { auth } from '/src/firebase/firebaseConfig'
+
+function useAuthState() {
+  const [user, setUser] = useState(null)
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
+      setUser(currentUser)
+    })
+
+    return () => unsubscribe()
+  }, [])
+
+  return user
+}
+
+export default useAuthState
