+++ src/components/Navbar/UserStatus.jsx
@@ -4,22 +4,23 @@ import styled from 'styled-components'
 import SettingIcon from '@assets/icons/setting.svg'
 import AlertIcon from '@assets/icons/alert.svg'
 import ProfileIcon from '@assets/icons/profile.svg'
-import { useNavigate } from 'react-router-dom'
-import { auth } from '/src/firebase/firebaseConfig'
+// import { useNavigate } from 'react-router-dom'
+// import { auth } from '/src/firebase/firebaseConfig'
+import useHandleLogout from '@components/Login/useHandleLogout'
 
 function UserStatus() {
-  const navigate = useNavigate()
-  const handleLogout = () => {
-    auth
-      .signOut()
-      .then(() => {
-        sessionStorage.removeItem('user')
-        navigate('/signin')
-      })
-      .catch((error) => {
-        console.error(error)
-      })
-  }
+  const handleLogout = useHandleLogout()
+  // const navigate = useNavigate()
+  // const handleLogout = () => {
+  //   auth
+  //     .signOut()
+  //     .then(() => {
+  //       navigate('/signin')
+  //     })
+  //     .catch((error) => {
+  //       console.error(error)
+  //     })
+  // }
   return (
     <Container $gap="20px" $justify="flex-start">
       <Flex $gap="10px">
