+++ src/pages/signin/signin-page.jsx
@@ -101,9 +101,8 @@ function SigninPage() {
 
   const handleLogin = async () => {
     try {
-      const curUser = await signInWithEmailAndPassword(auth, email, password)
-      console.log('login successed')
-      sessionStorage.setItem('user', JSON.stringify(curUser.user))
+      const user = await signInWithEmailAndPassword(auth, email, password)
+      console.log(user)
       navigate('/')
     } catch (error) {
       setError('유효한 아이디, 비밀번호를 입력해주세요!')
